dtype bug

+++ src/models/classifier_head.py
@@ -14,7 +14,7 @@ class PooledClassifier(nn.Module):
         self.linear = nn.Linear(hidden_size, num_labels)
 
     def forward(self, last_hidden_state: torch.Tensor, attention_mask: torch.Tensor) -> torch.Tensor:
-        mask = attention_mask.unsqueeze(-1).float()
+        mask = attention_mask.unsqueeze(-1).to(dtype=last_hidden_state.dtype)
         pooled = (last_hidden_state * mask).sum(dim=1) / mask.sum(dim=1).clamp(min=1e-9)
         return self.linear(self.dropout(pooled))
